agents/nyc_data/requirements.txt
google-genai>=0.8.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
